--- src/src/main/java/com/hcmus/mentor/backend/domain/UserGender.java
@@ -1,6 +1,0 @@
-package com.hcmus.mentor.backend.domain;
-
-public enum UserGender {
-    FEMALE,
-    MALE
-}
