+++ src/src/main/java/com/hcmus/mentor/backend/domain/constant/NotificationType.java
@@ -1,0 +1,18 @@
+package com.hcmus.mentor.backend.domain.constant;
+
+public enum NotificationType {
+    SYSTEM,
+    NEW_COMMENT,
+    NEW_MESSAGE,
+    NEW_TASK,
+    NEW_MEETING,
+    NEW_REACTION,
+    NEW_IMAGE_MESSAGE,
+    NEW_FILE_MESSAGE,
+    UPDATE_MEETING,
+    RESCHEDULE_MEETING,
+    NEW_VOTE,
+    PIN_MESSAGE,
+    UNPIN_MESSAGE,
+    NEED_RESPONSE
+}
