+++ src/src/main/java/com/hcmus/mentor/backend/domain/constant/ReminderType.java
@@ -1,4 +1,4 @@
-package com.hcmus.mentor.backend.domain;
+package com.hcmus.mentor.backend.domain.constant;
 
 public enum ReminderType {
     TASK,
