+++ src/src/main/java/com/hcmus/mentor/backend/domain/constant/UserGender.java
@@ -1,0 +1,6 @@
+package com.hcmus.mentor.backend.domain.constant;
+
+public enum UserGender {
+    FEMALE,
+    MALE
+}
